+++ .gitignore
@@ -156,3 +156,4 @@ cython_debug/
 *.model
 .DS_Store
 checkpoints/
+.vscode/
