+++ src/chatbot/__init__.py
@@ -1,1 +1,1 @@
-from .chain import DummyChatBot, LitGPTChatBot, ServerChatBot
+from .conversation_chain import DummyChatBot, LitGPTChatBot, ServerChatBot
